Enhanced Print Generation Speed
